所持スキルカードダイアログの表示用情報を生成 / 表示関連処理を大リファクタリング
