+++ src/index.test.ts
@@ -1,75 +1,27 @@
 import type {
   Card,
-  CardEnhancement,
   CardInProduction,
   Effect,
   Lesson,
-  LessonDisplay,
   LessonGamePlay,
   LessonUpdateQuery,
-  Modifier,
 } from "./index";
 import { getCardDataById } from "./data/cards";
 import { getProducerItemDataById } from "./data/producer-items";
 import {
-  createLessonDisplay,
   endLessonTurn,
-  getEncouragements,
   getNextHistoryResultIndex,
   initializeGamePlay,
   isLessonEnded,
   isTurnEnded,
   patchUpdates,
   playCard,
-  previewCardPlay,
   startLessonTurn,
 } from "./index";
 import { activateEffect as activateEffect_ } from "./lesson-mutation";
 import { createLessonGamePlayForTest } from "./test-utils";
 import { getIdolDataById } from "./data/idols";
 
-describe("getEncouragements", () => {
-  const testCases: Array<{
-    args: Parameters<typeof getEncouragements>;
-    expected: ReturnType<typeof getEncouragements>;
-    name: string;
-  }> = [
-    {
-      name: "descriptionの条件と効果の間に読点がない",
-      args: [
-        {
-          initialLesson: {
-            encouragements: [
-              {
-                turnNumber: 1,
-                effect: {
-                  kind: "perform",
-                  vitality: { value: 1 },
-                  condition: {
-                    kind: "countModifier",
-                    modifierKind: "focus",
-                    range: { min: 3 },
-                  },
-                },
-              },
-            ],
-          },
-          updates: [] as LessonUpdateQuery[],
-        } as LessonGamePlay,
-      ],
-      expected: [
-        {
-          turnNumber: 1,
-          effect: expect.any(Object),
-          description: "{{集中}}が3以上の場合{{元気}}+1",
-        },
-      ],
-    },
-  ];
-  test.each(testCases)("$name", ({ args, expected }) => {
-    expect(getEncouragements(...args)).toStrictEqual(expected);
-  });
-});
 describe("initializeGamePlay", () => {
   const testCases: Array<{
     args: Parameters<typeof initializeGamePlay>;
@@ -192,277 +144,6 @@ describe("initializeGamePlay", () => {
   ];
   test.each(testCases)("$name", ({ args, expected }) => {
     expect(initializeGamePlay(...args)).toMatchObject(expected);
-  });
-});
-describe("createLessonDisplay", () => {
-  const testCases: Array<{
-    args: Parameters<typeof createLessonDisplay>;
-    expected: ReturnType<typeof createLessonDisplay>;
-    name: string;
-  }> = [
-    {
-      name: "hand - 概ね動作する",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest({
-            deck: [
-              {
-                id: "c1",
-                data: getCardDataById("apirunokihon"),
-                enhanced: false,
-              },
-              {
-                id: "c2",
-                data: getCardDataById("hyogennokihon"),
-                enhanced: true,
-              },
-            ],
-          });
-          gamePlay.initialLesson.hand = ["c1", "c2"];
-          gamePlay.initialLesson.cards = gamePlay.initialLesson.cards.map(
-            (card) => {
-              return card.id === "c2"
-                ? {
-                    ...card,
-                    enhancements: [
-                      ...card.enhancements,
-                      { kind: "lessonSupport" },
-                    ],
-                  }
-                : card;
-            },
-          );
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        hand: [
-          {
-            name: "アピールの基本",
-            enhancements: [] as CardEnhancement[],
-          },
-          {
-            name: "表現の基本++",
-            enhancements: [{ kind: "original" }, { kind: "lessonSupport" }],
-          },
-        ],
-      } as LessonDisplay,
-    },
-    {
-      name: 'producerItems - 概ね動作する | 強化済みの名称には、"+"を末尾へ付与する',
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest({
-            producerItems: [
-              {
-                id: "p1",
-                data: getProducerItemDataById("saigononatsunoomoide"),
-                enhanced: true,
-              },
-            ],
-          });
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        producerItems: [
-          {
-            id: "p1",
-            name: "最後の夏の思い出+",
-            description: [
-              "ターン開始時{{集中}}が3以上の場合、{{元気}}+14",
-              "（レッスン内1回）",
-            ].join("\n"),
-            activationCount: 0,
-            remainingTimes: 1,
-          },
-        ],
-      } as LessonDisplay,
-    },
-    {
-      name: "modifiers - 状態修正が存在する時、それを含む配列を返す",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest();
-          gamePlay.initialLesson.idol.modifiers = [
-            { kind: "focus", amount: 1, id: "m1" },
-          ];
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        modifiers: [
-          {
-            id: "m1",
-            kind: "focus",
-            amount: 1,
-            label: "集中",
-            description: expect.any(String),
-          },
-        ],
-      } as LessonDisplay,
-    },
-    {
-      name: "modifiers - 状態修正が存在しない時、空配列を返す",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest();
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        modifiers: [] as Modifier[],
-      } as LessonDisplay,
-    },
-    {
-      name: "turns - 概ね動作する",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest();
-          gamePlay.initialLesson.turns = ["vocal", "visual", "dance"];
-          gamePlay.initialLesson.encouragements = [
-            {
-              turnNumber: 3,
-              effect: { kind: "perform", vitality: { value: 1 } },
-            },
-          ];
-          gamePlay.initialLesson.turnNumber = 2;
-          gamePlay.initialLesson.remainingTurns = 1;
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        turns: [
-          {
-            turnNumber: 1,
-            turnNumberDiff: -1,
-            idolParameterKind: "vocal",
-            idolParameterLabel: "ボーカル",
-          },
-          {
-            turnNumber: 2,
-            turnNumberDiff: 0,
-            idolParameterKind: "visual",
-            idolParameterLabel: "ビジュアル",
-          },
-          {
-            turnNumber: 3,
-            turnNumberDiff: 1,
-            idolParameterKind: "dance",
-            idolParameterLabel: "ダンス",
-            encouragement: {
-              turnNumber: 3,
-              effect: expect.any(Object),
-              description: "{{元気}}+1",
-            },
-          },
-          {
-            turnNumber: 4,
-            turnNumberDiff: 2,
-            idolParameterKind: "dance",
-            idolParameterLabel: "ダンス",
-          },
-        ],
-      } as LessonDisplay,
-    },
-    {
-      name: "scoreBonus - 概ね動作する",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest();
-          gamePlay.initialLesson.turns = ["vocal"];
-          gamePlay.initialLesson.turnNumber = 1;
-          gamePlay.initialLesson.idol.scoreBonus = {
-            vocal: 300,
-            visual: 200,
-            dance: 100,
-          };
-          return gamePlay;
-        })(),
-      ],
-      expected: {
-        scoreBonus: 300,
-      } as LessonDisplay,
-    },
-  ];
-  test.each(testCases)("$name", ({ args, expected }) => {
-    expect(createLessonDisplay(...args)).toMatchObject(expected);
-  });
-});
-describe("previewCardPlay", () => {
-  const testCases: Array<{
-    args: Parameters<typeof previewCardPlay>;
-    expected: ReturnType<typeof previewCardPlay>;
-    name: string;
-  }> = [
-    {
-      name: "概ね動作する",
-      args: [
-        (() => {
-          const gamePlay = createLessonGamePlayForTest({
-            deck: [
-              {
-                id: "c1",
-                data: getCardDataById("jonetsutan"),
-                enhanced: true,
-              },
-            ],
-            producerItems: [
-              {
-                id: "p1",
-                data: getProducerItemDataById("itsumonomeikupochi"),
-                enhanced: false,
-              },
-            ],
-          });
-          gamePlay.initialLesson.hand = ["c1"];
-          gamePlay.initialLesson.idol.modifiers = [
-            { kind: "halfLifeConsumption", duration: 1, id: "m1" },
-          ];
-          return gamePlay;
-        })(),
-        0,
-      ],
-      expected: {
-        cardName: "情熱ターン+",
-        cardDescription: ["パラメータ+18", "{{集中}}+4"].join("\n"),
-        // スキルカードのプレビューには、消費体力減少効果は反映されていない
-        cardCost: { kind: "normal", value: 6 },
-        // 「いつものメイクポーチ」は、本来発動するはずだが、プレビューなので発動していない
-        updates: [
-          // 差分には、消費体力減少効果が反映されている
-          {
-            kind: "life",
-            actual: -3,
-            max: -3,
-            reason: expect.any(Object),
-          },
-          {
-            kind: "score",
-            actual: 18,
-            max: 18,
-            reason: expect.any(Object),
-          },
-          {
-            kind: "modifier",
-            actual: {
-              kind: "focus",
-              amount: 4,
-              id: expect.any(String),
-            },
-            max: {
-              kind: "focus",
-              amount: 4,
-              id: expect.any(String),
-            },
-            reason: expect.any(Object),
-          },
-        ],
-      },
-    },
-  ];
-  test.each(testCases)("$name", ({ args, expected }) => {
-    expect(previewCardPlay(...args)).toStrictEqual(expected);
   });
 });
